+++ .gitignore
@@ -1,2 +1,2 @@
-/node_modules
-records.json
+/server/node_modules
+/server/records.json
